refactor(server): extract HF token resolution to hf_token.py
The 40-line `resolve_hf_token` block sat at the top of server.py with
five env-var / file-path fallbacks; it's process-launch infrastructure,
not server logic. Move it to its own module behind a single
`apply_resolved_token()` entry point that resolves and stamps HF_TOKEN
into os.environ before the heavy imports run.

Drops the last "side effect at module-import time" idiom from
server.py; the call is now an explicit one-liner that file-readers
can grep for.

Refs #78.
